+++ CHANGELOG.md
@@ -7,7 +7,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased] 
 
-- When adding to existing graph the source graph IRI is added to the label
+- When adding to existing graph the source graph IRI is appended to the label. If the original label does not conform
+to the format "Shapes for: [IRI], [IRI2], ..." it is prefixed with "Previous label:" and a new label is added with the
+new data graph IRI.
 
 ### Added